refactor: remove rustfs owner compat bridges
