+++ rustfs/src/admin/handlers/replication.rs
@@ -12,15 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::storage_compat::StorageError;
-use super::super::storage_compat::bucket_target_sys::{BucketTargetError, BucketTargetSys};
-use super::super::storage_compat::global_rustfs_port;
-use super::super::storage_compat::metadata::BUCKET_TARGETS_FILE;
-use super::super::storage_compat::metadata_sys;
-use super::super::storage_compat::metadata_sys::get_replication_config;
-use super::super::storage_compat::replication::BucketStats;
-use super::super::storage_compat::replication::GLOBAL_REPLICATION_STATS;
-use super::super::storage_compat::target::BucketTarget;
+use super::super::StorageError;
+use super::super::bucket_target_sys::{BucketTargetError, BucketTargetSys};
+use super::super::global_rustfs_port;
+use super::super::metadata::BUCKET_TARGETS_FILE;
+use super::super::metadata_sys;
+use super::super::metadata_sys::get_replication_config;
+use super::super::replication::BucketStats;
+use super::super::replication::GLOBAL_REPLICATION_STATS;
+use super::super::target::BucketTarget;
 use crate::admin::auth::validate_admin_request;
 use crate::admin::handlers::site_replication::site_replication_peer_deployment_id_for_endpoint;
 use crate::admin::router::{AdminOperation, Operation, S3Router};
