+++ rustfs/src/admin/router.rs
@@ -12,23 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::storage_compat::GLOBAL_BOOT_TIME;
-use super::storage_compat::PeerRestClient;
-use super::storage_compat::bandwidth::monitor::BandwidthDetails;
-use super::storage_compat::bucket_target_sys::{
-    BucketTargetSys, PutObjectOptions, RemoveObjectOptions, S3ClientError, TargetClient,
-};
-use super::storage_compat::get_global_notification_sys;
-use super::storage_compat::metadata::BUCKET_TARGETS_FILE;
-use super::storage_compat::metadata_sys;
-use super::storage_compat::read_admin_config_without_migrate;
-use super::storage_compat::replication::{
+use super::GLOBAL_BOOT_TIME;
+use super::PeerRestClient;
+use super::bandwidth::monitor::BandwidthDetails;
+use super::bucket_target_sys::{BucketTargetSys, PutObjectOptions, RemoveObjectOptions, S3ClientError, TargetClient};
+use super::get_global_notification_sys;
+use super::metadata::BUCKET_TARGETS_FILE;
+use super::metadata_sys;
+use super::read_admin_config_without_migrate;
+use super::replication::{
     BucketReplicationResyncStatus, BucketStats, GLOBAL_REPLICATION_STATS, ObjectOpts, ResyncOpts, get_global_replication_pool,
 };
-use super::storage_compat::target::{BucketTarget, BucketTargetType, BucketTargets};
-use super::storage_compat::versioning_sys::BucketVersioningSys;
-use super::storage_compat::{AdminReplicationConfigExt as _, AdminVersioningConfigExt as _};
-use super::storage_compat::{get_global_bucket_monitor, get_global_deployment_id, get_global_region};
+use super::target::{BucketTarget, BucketTargetType, BucketTargets};
+use super::versioning_sys::BucketVersioningSys;
+use super::{AdminReplicationConfigExt as _, AdminVersioningConfigExt as _};
+use super::{get_global_bucket_monitor, get_global_deployment_id, get_global_region};
 use crate::admin::console::{is_console_path, make_console_server};
 use crate::admin::handlers::oidc::is_oidc_path;
 use crate::app::context::resolve_object_store_handle;
@@ -1419,7 +1417,7 @@ async fn ensure_replication_bucket_exists(bucket: &str) -> S3Result<()> {
 async fn ensure_replication_config_exists(bucket: &str) -> S3Result<()> {
     match metadata_sys::get_replication_config(bucket).await {
         Ok(_) => Ok(()),
-        Err(super::storage_compat::StorageError::ConfigNotFound) => Err(s3_error!(ReplicationConfigurationNotFoundError)),
+        Err(super::StorageError::ConfigNotFound) => Err(s3_error!(ReplicationConfigurationNotFoundError)),
         Err(err) => Err(ApiError::from(err).into()),
     }
 }
@@ -1944,7 +1942,7 @@ async fn resolve_replication_target_client(bucket: &str, target: &BucketTarget) 
 
 fn build_replication_probe_put_options(now: OffsetDateTime) -> PutObjectOptions {
     PutObjectOptions {
-        internal: super::storage_compat::bucket_target_sys::AdvancedPutOptions {
+        internal: super::bucket_target_sys::AdvancedPutOptions {
             source_version_id: Uuid::new_v4().to_string(),
             replication_status: ReplicationStatusType::Replica,
             source_mtime: now,
@@ -2059,7 +2057,7 @@ async fn source_bucket_requires_object_lock(bucket: &str) -> S3Result<bool> {
             .object_lock_enabled
             .as_ref()
             .is_some_and(|state| state.as_str() == s3s::dto::ObjectLockEnabled::ENABLED)),
-        Err(super::storage_compat::StorageError::ConfigNotFound) => Ok(false),
+        Err(super::StorageError::ConfigNotFound) => Ok(false),
         Err(err) => Err(ApiError::from(err).into()),
     }
 }
@@ -2773,7 +2771,7 @@ mod tests {
     #[test]
     fn apply_replication_reset_to_targets_updates_matching_target() {
         let mut targets = BucketTargets {
-            targets: vec![super::super::storage_compat::target::BucketTarget {
+            targets: vec![super::super::target::BucketTarget {
                 arn: "arn:target".to_string(),
                 ..Default::default()
             }],
@@ -2795,10 +2793,10 @@ mod tests {
         let mut status = BucketReplicationResyncStatus::new();
         status.targets_map.insert(
             "arn:z".to_string(),
-            super::super::storage_compat::replication::TargetReplicationResyncStatus {
+            super::super::replication::TargetReplicationResyncStatus {
                 resync_id: "rid-z".to_string(),
                 last_update: Some(datetime!(2025-01-03 00:00 UTC)),
-                resync_status: super::super::storage_compat::replication::ResyncStatusType::ResyncFailed,
+                resync_status: super::super::replication::ResyncStatusType::ResyncFailed,
                 failed_count: 2,
                 failed_size: 4,
                 bucket: "bucket-z".to_string(),
@@ -2808,10 +2806,10 @@ mod tests {
         );
         status.targets_map.insert(
             "arn:a".to_string(),
-            super::super::storage_compat::replication::TargetReplicationResyncStatus {
+            super::super::replication::TargetReplicationResyncStatus {
                 resync_id: "rid-a".to_string(),
                 last_update: Some(datetime!(2025-01-02 00:00 UTC)),
-                resync_status: super::super::storage_compat::replication::ResyncStatusType::ResyncCompleted,
+                resync_status: super::super::replication::ResyncStatusType::ResyncCompleted,
                 replicated_count: 3,
                 replicated_size: 9,
                 bucket: "bucket-a".to_string(),
@@ -2841,10 +2839,10 @@ mod tests {
         let mut status = BucketReplicationResyncStatus::new();
         status.targets_map.insert(
             "arn:z".to_string(),
-            super::super::storage_compat::replication::TargetReplicationResyncStatus {
+            super::super::replication::TargetReplicationResyncStatus {
                 resync_id: "rid-z".to_string(),
                 last_update: Some(datetime!(2025-02-03 00:00 UTC)),
-                resync_status: super::super::storage_compat::replication::ResyncStatusType::ResyncFailed,
+                resync_status: super::super::replication::ResyncStatusType::ResyncFailed,
                 failed_count: 2,
                 failed_size: 4,
                 bucket: "bucket-z".to_string(),
@@ -2854,10 +2852,10 @@ mod tests {
         );
         status.targets_map.insert(
             "arn:a".to_string(),
-            super::super::storage_compat::replication::TargetReplicationResyncStatus {
+            super::super::replication::TargetReplicationResyncStatus {
                 resync_id: "rid-a".to_string(),
                 last_update: Some(datetime!(2025-02-02 00:00 UTC)),
-                resync_status: super::super::storage_compat::replication::ResyncStatusType::ResyncCompleted,
+                resync_status: super::super::replication::ResyncStatusType::ResyncCompleted,
                 replicated_count: 3,
                 replicated_size: 9,
                 bucket: "bucket-a".to_string(),
