+++ rustfs/src/app/bucket_usecase.rs
@@ -14,12 +14,12 @@
 
 //! Bucket application use-case contracts.
 
-use super::storage_compat::ECStore;
-use super::storage_compat::StorageError;
-use super::storage_compat::get_global_notification_sys;
-use super::storage_compat::object_api_utils::to_s3s_etag;
-use super::storage_compat::{AppObjectLockConfigExt as _, AppVersioningConfigExt as _};
-use super::storage_compat::{
+use super::ECStore;
+use super::StorageError;
+use super::get_global_notification_sys;
+use super::object_api_utils::to_s3s_etag;
+use super::{AppObjectLockConfigExt as _, AppVersioningConfigExt as _};
+use super::{
     bucket_target_sys::BucketTargetSys,
     lifecycle::bucket_lifecycle_ops::{
         enqueue_expiry_for_existing_objects, enqueue_transition_for_existing_objects, validate_lifecycle_config,
@@ -685,7 +685,7 @@ fn assign_lifecycle_rule_ids(rules: &mut [LifecycleRule]) {
     }
 }
 
-fn validate_lifecycle_rule_status(rules: &[LifecycleRule]) -> Result<(), &'static str> {
+fn validate_lifecycle_rule_status(rules: &[LifecycleRule]) -> std::result::Result<(), &'static str> {
     for rule in rules {
         if rule.status != ExpirationStatus::from_static(ExpirationStatus::ENABLED)
             && rule.status != ExpirationStatus::from_static(ExpirationStatus::DISABLED)
@@ -2285,7 +2285,7 @@ mod tests {
         BucketTargets {
             targets: arns
                 .iter()
-                .map(|arn| super::super::storage_compat::target::BucketTarget {
+                .map(|arn| super::super::target::BucketTarget {
                     arn: (*arn).to_string(),
                     target_type: BucketTargetType::ReplicationService,
                     ..Default::default()
