+++ rustfs/src/app/context/handles.rs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::storage_compat::EndpointServerPools;
-use super::super::storage_compat::TierConfigMgr;
-use super::super::storage_compat::metadata_sys::{BucketMetadataSys, get_global_bucket_metadata_sys};
-use super::super::storage_compat::{get_global_endpoints_opt, get_global_region, get_global_tier_config_mgr};
+use super::super::EndpointServerPools;
+use super::super::TierConfigMgr;
+use super::super::metadata_sys::{BucketMetadataSys, get_global_bucket_metadata_sys};
+use super::super::{get_global_endpoints_opt, get_global_region, get_global_tier_config_mgr};
 use super::interfaces::{
     BucketMetadataInterface, BufferConfigInterface, EndpointsInterface, IamInterface, KmsInterface, KmsRuntimeInterface,
     NotifyInterface, RegionInterface, ServerConfigInterface, TierConfigInterface,
