+++ rustfs/src/app/context/startup.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::super::storage_compat::ECStore;
+use super::super::ECStore;
 use super::global::{AppContext, get_global_app_context, init_global_app_context};
 use rustfs_kms::KmsServiceManager;
 use std::io::{Error, Result};
