+++ rustfs/src/storage/ecfs_test.rs
@@ -14,9 +14,7 @@
 
 #[cfg(test)]
 mod tests {
-    use super::super::storage_compat::{
-        BucketMetadata, DEFAULT_READ_BUFFER_SIZE, get_global_bucket_metadata_sys, set_bucket_metadata,
-    };
+    use super::super::{BucketMetadata, DEFAULT_READ_BUFFER_SIZE, get_global_bucket_metadata_sys, set_bucket_metadata};
     use crate::config::WorkloadProfile;
     use crate::server::cors;
     use crate::storage::ecfs::{FS, validate_object_lock_configuration_input};
@@ -941,8 +939,8 @@ mod tests {
 
     #[tokio::test]
     async fn test_validate_bucket_object_lock_enabled() {
-        use super::super::storage_compat::bucket_metadata_sys_initialized;
-        use super::super::storage_compat::set_bucket_metadata;
+        use super::super::bucket_metadata_sys_initialized;
+        use super::super::set_bucket_metadata;
         use s3s::dto::{ObjectLockConfiguration, ObjectLockEnabled};
         use time::OffsetDateTime;
 
@@ -1781,7 +1779,7 @@ mod tests {
     /// with a single-element vec value, matching the format expected by policy evaluation.
     #[test]
     fn test_object_tag_condition_key_format() {
-        use super::super::storage_compat::decode_tags_to_map;
+        use super::super::decode_tags_to_map;
         use std::collections::HashMap;
 
         let tags_str = "security=public&project=webapp&env=prod";
